calculs polygone

+++ src/model/ShapeComposant.java
@@ -2,8 +2,8 @@ package model;
 
 public interface ShapeComposant {
 	public void translation();
-	public int getXpos();
-	public void setXpos(int xpos);
-	public int getYpos();
-	public void setYpos(int ypos);
+	public double getXpos();
+	public void setXpos(double xpos);
+	public double getYpos();
+	public void setYpos(double ypos);
 }
